+++ R/add_predictors.R
@@ -60,6 +60,7 @@ NULL
 #' * \code{'interaction'} - Add interacting predictors. Interactions need to be specified (\code{"int_variables"})!
 #' * \code{'thresh'} - Add threshold derivate predictors.
 #' * \code{'hinge'} - Add hinge derivate predictors.
+#' * \code{'kmeans'} - Add k-means derived factors.
 #' * \code{'bin'} - Add predictors binned by their percentiles.
 #'
 #' @note
@@ -130,7 +131,7 @@ methods::setMethod(
     assertthat::assert_that(inherits(x, "BiodiversityDistribution"),
                             is.Raster(env),
                             all(transform == 'none') || all( transform %in% c('pca', 'scale', 'norm', 'windsor') ),
-                            all(derivates == 'none') || all( derivates %in% c('thresh', 'hinge', 'quadratic', 'bin', 'interaction') ),
+                            all(derivates == 'none') || all( derivates %in% c('thresh', 'hinge', 'quadratic', 'bin', 'kmeans', 'interaction') ),
                             is.vector(derivate_knots) || is.numeric(derivate_knots),
                             is.null(names) || assertthat::is.scalar(names) || is.vector(names),
                             is.logical(explode_factors),
@@ -753,7 +754,7 @@ methods::setMethod(
     # names = names = NULL; transform = 'none'; derivates = 'none'; derivate_knots = 4; int_variables = NULL;harmonize_na = FALSE; state = NULL
     # Try and match transform and derivatives arguments
     transform <- match.arg(transform, c('none','pca', 'scale', 'norm', 'windsor', 'percentile'), several.ok = FALSE) # Several ok set to FALSE as states are not working otherwise
-    derivates <- match.arg(derivates, c('none','thresh', 'hinge', 'quadratic', 'bin', 'interaction'), several.ok = TRUE)
+    derivates <- match.arg(derivates, c('none','thresh', 'hinge', 'quadratic', 'bin', 'kmeans', 'interaction'), several.ok = TRUE)
 
     assertthat::validate_that(inherits(env,'stars'), msg = 'Projection rasters need to be stars stack!')
     assertthat::assert_that(inherits(x, "BiodiversityScenario"),
@@ -767,8 +768,8 @@ methods::setMethod(
     assertthat::validate_that(length(env) >= 1)
 
     # Get model object
-    obj <- x$get_model()
-    assertthat::assert_that(!(is.null(obj) || is.Waiver(obj)),
+    obj <- x$get_model(copy = TRUE)
+    assertthat::assert_that(!(isFALSE(obj) || is.Waiver(obj)),
                             msg = "No model object found in scenario?")
     model <- obj$model
 
@@ -856,7 +857,7 @@ methods::setMethod(
       # Get variable names
       varn <- obj$get_coefficients()[,1]
       # Are there any derivates present in the coefficients?
-      if(any( length( grep("hinge_|bin_|quadratic_|thresh_|interaction_", varn ) ) > 0 )){
+      if(any( length( grep("hinge_|bin_|kmeans_|quadratic_|thresh_|interaction_", varn ) ) > 0 )){
         if(getOption('ibis.setupmessages', default = TRUE)) myLog('[Setup]','green','Creating predictor derivates...')
         for(dd in derivates){
           if(any(grep(dd, varn))){